refactor: update keydown folder structure
